fix: 상품 도메인 Clean + Layered Architecture 적용
