+++ src/main/java/kr/hhplus/be/server/domain/product/Product.java
@@ -1,0 +1,10 @@
+package kr.hhplus.be.server.domain.product;
+
+public record Product(
+    long id,
+    String name,
+    int price,
+    int stock
+) {
+
+}
