+++ src/main/java/kr/hhplus/be/server/infrastructure/product/ProductRepository.java
@@ -1,0 +1,5 @@
+package kr.hhplus.be.server.infrastructure.product;
+
+public interface ProductRepository {
+
+}
