+++ src/main/java/kr/hhplus/be/server/infrastructure/product/ProductRepositoryImpl.java
@@ -1,0 +1,5 @@
+package kr.hhplus.be.server.infrastructure.product;
+
+public class ProductRepositoryImpl {
+
+}
