+++ src/main/java/kr/hhplus/be/server/interfaces/product/ProductApi.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.product;
+package kr.hhplus.be.server.interfaces.product;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -9,6 +9,7 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import java.util.List;
+import kr.hhplus.be.server.domain.product.Product;
 import org.springframework.web.ErrorResponse;
 
 @Tag(name = "Product", description = "상품 관련 API")
@@ -21,7 +22,7 @@ public interface ProductApi {
         @ApiResponse(responseCode = "500", description = "서버 에러",
             content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
     })
-    List<Product> getAllProducts();
+    List<ProductResponse.Product> getAllProducts();
 
     @Operation(summary = "상품 상세 조회")
     @Parameter(name = "productId", description = "상품 ID", required = true)
@@ -31,7 +32,7 @@ public interface ProductApi {
         @ApiResponse(responseCode = "500", description = "서버 에러",
             content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
     })
-    Product getProduct(long productId);
+    ProductResponse.Product getProduct(long productId);
 
     @Operation(summary = "Top 5 인기 상품 조회")
     @ApiResponses({
@@ -40,6 +41,6 @@ public interface ProductApi {
         @ApiResponse(responseCode = "500", description = "서버 에러",
             content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
     })
-    List<Product> getTop5Products();
+    List<ProductResponse.Product> getTop5Products();
 
 }
